Move weight bitpacking into its own converter pass.
This is a prerequisite for #391.
