+++ larq_compute_engine/mlir/tests/optimize.mlir
@@ -151,7 +151,7 @@ func @tranform_negative_multipliers(%arg0: tensor<256x32x32x3xf32>, %arg1: tenso
   %0 = "tf.LceBconv2d"(%arg0, %filter, %multiplier, %arg1) {channels_in = 3 : i32, dilation_height_factor = 1 : i32, dilation_width_factor = 1 : i32, fused_activation_function = "NONE", padding = "VALID", stride_height = 1 : i32, stride_width = 1 : i32} : (tensor<256x32x32x3xf32>, tensor<16x3x3x3xf32>, tensor<16xf32>, tensor<16xf32>) -> tensor<256x30x30x16xf32>
   return %0 : tensor<256x30x30x16xf32>
 
-  // CHECK: %cst = constant dense<0> : tensor<16x3x3x1xi32>
+  // CHECK: %cst = constant dense<1.000000e+00> : tensor<16x3x3x3xf32>
   // CHECK: %cst_0 = constant dense<3.000000e+00> : tensor<16xf32>
   // CHECK: %0 = "tf.LceBconv2d"(%arg0, %cst, %cst_0, %arg1)
   // CHECK-NEXT: return %0
@@ -165,19 +165,8 @@ func @do_not_change_multiplier_if_fused_activation(%arg0: tensor<256x32x32x3xf32
   return %0 : tensor<256x30x30x16xf32>
 
   // CHECK: %cst = constant dense<-3.000000e+00> : tensor<16xf32>
-  // CHECK: %cst_0 = constant dense<0> : tensor<16x3x3x1xi32>
+  // CHECK: %cst_0 = constant dense<1.000000e+00> : tensor<16x3x3x3xf32>
   // CHECK: %0 = "tf.LceBconv2d"(%arg0, %cst_0, %cst, %arg1)
-  // CHECK-NEXT: return %0
-}
-
-// CHECK-LABEL: @bitpack_bconv2d_filters
-func @bitpack_bconv2d_filters(%arg0: tensor<256x32x32x3xf32>, %arg1: tensor<16xf32>, %arg2: tensor<16xf32>) -> tensor<256x30x30x16xf32> {
-  %cst = constant dense<1.0> : tensor<16x3x3x3xf32>
-  %0 = "tf.LceBconv2d"(%arg0, %cst, %arg1, %arg2) {channels_in = 3 : i32, dilation_height_factor = 1 : i32, dilation_width_factor = 1 : i32, fused_activation_function = "NONE", padding = "VALID", stride_height = 1 : i32, stride_width = 1 : i32} : (tensor<256x32x32x3xf32>, tensor<16x3x3x3xf32>, tensor<16xf32>, tensor<16xf32>) -> tensor<256x30x30x16xf32>
-  return %0 : tensor<256x30x30x16xf32>
-
-  // CHECK: %cst = constant dense<0> : tensor<16x3x3x1xi32>
-  // CHECK: %0 = "tf.LceBconv2d"(%arg0, %cst, %arg1, %arg2) {channels_in = 3 : i32, dilation_height_factor = 1 : i32, dilation_width_factor = 1 : i32, fused_activation_function = "NONE", pad_values = 0 : i32, padding = "VALID", stride_height = 1 : i32, stride_width = 1 : i32} : (tensor<256x32x32x3xf32>, tensor<16x3x3x1xi32>, tensor<16xf32>, tensor<16xf32>) -> tensor<256x30x30x16xf32>
   // CHECK-NEXT: return %0
 }
